@@ -1,12 +1,12 @@
 name: webfeed
-version: 0.6.0
+version: 0.7.0
 description: webfeed is a dart package for parsing RSS and Atom feeds. Media, DublinCore, iTunes, Syndication namespaces are also supported.
 homepage: https://github.com/witochandra/webfeed
 environment:
-  sdk: ">=2.0.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 dependencies:
   xml: "^5.0.2"
-  intl: "^0.17.0-nullsafety.2"
+  intl: "^0.17.0"
 dev_dependencies:
   test: ^1.3.0
-  http: "^0.11.3+16"
+  http: "^0.13.0"
