@@ -3,27 +3,28 @@ import 'package:webfeed/domain/itunes/itunes_episode_type.dart';
 import 'package:webfeed/domain/itunes/itunes_image.dart';
 import 'package:webfeed/domain/itunes/itunes_owner.dart';
 import 'package:webfeed/domain/itunes/itunes_type.dart';
+import 'package:webfeed/util/iterable.dart';
 import 'package:webfeed/util/xml.dart';
 import 'package:xml/xml.dart';
 
 class Itunes {
-  final String author;
-  final String summary;
-  final bool explicit;
-  final String title;
-  final String subtitle;
-  final ItunesOwner owner;
-  final List<String> keywords;
-  final ItunesImage image;
-  final List<ItunesCategory> categories;
-  final ItunesType type;
-  final String newFeedUrl;
-  final bool block;
-  final bool complete;
-  final int episode;
-  final int season;
-  final Duration duration;
-  final ItunesEpisodeType episodeType;
+  final String? author;
+  final String? summary;
+  final bool? explicit;
+  final String? title;
+  final String? subtitle;
+  final ItunesOwner? owner;
+  final List<String>? keywords;
+  final ItunesImage? image;
+  final List<ItunesCategory>? categories;
+  final ItunesType? type;
+  final String? newFeedUrl;
+  final bool? block;
+  final bool? complete;
+  final int? episode;
+  final int? season;
+  final Duration? duration;
+  final ItunesEpisodeType? episodeType;
 
   Itunes({
     this.author,
@@ -46,38 +47,52 @@ class Itunes {
   });
 
   factory Itunes.parse(XmlElement element) {
-    if (element == null) {
-      return null;
-    }
-    var episodeStr = findFirstElement(element, 'itunes:episode')?.text;
-    var seasonStr = findFirstElement(element, 'itunes:season')?.text;
-    var durationStr = findFirstElement(element, 'itunes:duration')?.text;
+    final episodeStr =
+        element.findElements('itunes:episode').firstOrNull?.text ?? '';
+    final seasonStr =
+        element.findElements('itunes:season').firstOrNull?.text ?? '';
+    final durationStr =
+        element.findElements('itunes:duration').firstOrNull?.text ?? '';
     return Itunes(
-      author: findFirstElement(element, 'itunes:author')?.text,
-      summary: findFirstElement(element, 'itunes:summary')?.text,
+      author: element.findElements('itunes:author').firstOrNull?.text,
+      summary: element.findElements('itunes:summary').firstOrNull?.text,
       explicit: parseBoolLiteral(element, 'itunes:explicit'),
-      title: findFirstElement(element, 'itunes:title')?.text,
-      subtitle: findFirstElement(element, 'itunes:subtitle')?.text,
-      owner: ItunesOwner.parse(findFirstElement(element, 'itunes:owner')),
-      keywords: findFirstElement(element, 'itunes:keywords')
+      title: element.findElements('itunes:title').firstOrNull?.text,
+      subtitle: element.findElements('itunes:subtitle').firstOrNull?.text,
+      owner: element
+          .findElements('itunes:owner')
+          .map((e) => ItunesOwner.parse(e))
+          .firstOrNull,
+      keywords: element
+              .findElements('itunes:keywords')
+              .firstOrNull
               ?.text
-              ?.split(',')
-              ?.map((keyword) => keyword.trim())
-              ?.toList() ??
+              .split(',')
+              .map((keyword) => keyword.trim())
+              .toList() ??
           [],
-      image: ItunesImage.parse(findFirstElement(element, 'itunes:image')),
-      categories: findElements(element, 'itunes:category')
+      image: element
+          .findElements('itunes:image')
+          .map((e) => ItunesImage.parse(e))
+          .firstOrNull,
+      categories: element
+          .findElements('itunes:category')
           .map((e) => ItunesCategory.parse(e))
           .toList(),
-      type: newItunesType(findFirstElement(element, 'itunes:type')),
-      newFeedUrl: findFirstElement(element, 'itunes:new-feed-url')?.text,
+      type: element
+          .findElements('itunes:type')
+          .map((e) => newItunesType(e))
+          .firstOrNull,
+      newFeedUrl: element.findElements('itunes:new-feed-url').firstOrNull?.text,
       block: parseBoolLiteral(element, 'itunes:block'),
       complete: parseBoolLiteral(element, 'itunes:complete'),
-      episode: episodeStr == null ? null : int.parse(episodeStr),
-      season: seasonStr == null ? null : int.parse(seasonStr),
-      duration: durationStr == null ? null : _parseDuration(durationStr),
-      episodeType:
-          newItunesEpisodeType(findFirstElement(element, 'itunes:episodeType')),
+      episode: episodeStr.isNotEmpty ? int.tryParse(episodeStr) : null,
+      season: seasonStr.isNotEmpty ? int.tryParse(seasonStr) : null,
+      duration: durationStr.isNotEmpty ? _parseDuration(durationStr) : null,
+      episodeType: element
+          .findElements('itunes:episodeType')
+          .map((e) => newItunesEpisodeType(e))
+          .firstOrNull,
     );
   }
 
@@ -87,12 +102,12 @@ class Itunes {
     var seconds = 0;
     var parts = s.split(s.contains(':') ? ':' : '.');
     if (parts.length > 2) {
-      hours = int.parse(parts[parts.length - 3]);
+      hours = int.tryParse(parts[parts.length - 3]) ?? 0;
     }
     if (parts.length > 1) {
-      minutes = int.parse(parts[parts.length - 2]);
+      minutes = int.tryParse(parts[parts.length - 2]) ?? 0;
     }
-    seconds = int.parse(parts[parts.length - 1]);
+    seconds = int.tryParse(parts[parts.length - 1]) ?? 0;
     return Duration(
       hours: hours,
       minutes: minutes,
